Refactor all of this code into an installable python module. Prune outdated code. Cleanup typos and code formatting, simplify file names, remove version numbers from file names.
